+++ src/lib/output/themes/default/assets/typedoc/components/Search.ts
@@ -260,7 +260,7 @@ function boldMatches(text: string, search: string) {
     const lowerText = text.toLocaleLowerCase();
     const lowerSearch = search.toLocaleLowerCase();
 
-    const parts = [];
+    const parts: string[] = [];
     let lastIndex = 0;
     let index = lowerText.indexOf(lowerSearch);
     while (index != -1) {