+++ src/test/behaviorTests.ts
@@ -499,4 +499,53 @@ export const behaviorTests: {
             "Single tag"
         );
     },
+
+    _searchCategoryBoosts(app) {
+        app.options.setValue("searchCategoryBoosts", {
+            Cat0: 0,
+            Cat1: 2.0,
+            Cat2: 1.5,
+            CatUnused: 999,
+        });
+    },
+    searchCategoryBoosts(project, logger) {
+        const a = query(project, "A");
+        const b = query(project, "B");
+        const c = query(project, "C");
+        equal(a.relevanceBoost, 3.0);
+        equal(b.relevanceBoost, 0.0);
+        equal(c.relevanceBoost, 2.0);
+        logger.expectMessage(
+            "warn: Not all categories specified in searchCategoryBoosts were used in the documentation." +
+                " The unused categories were:\n\tCatUnused"
+        );
+        logger.discardDebugMessages();
+        logger.expectNoOtherMessages();
+    },
+
+    _searchGroupBoosts(app) {
+        app.options.setValue("searchGroupBoosts", {
+            Group0: 0,
+            Group1: 2.0,
+            Group2: 1.5,
+            GroupUnused: 999,
+            Interfaces: 0.5,
+        });
+    },
+    searchGroupBoosts(project, logger) {
+        const a = query(project, "A");
+        const b = query(project, "B");
+        const c = query(project, "C");
+        const d = query(project, "D");
+        equal(a.relevanceBoost, 3.0);
+        equal(b.relevanceBoost, 0.0);
+        equal(c.relevanceBoost, 2.0);
+        equal(d.relevanceBoost, 0.5);
+        logger.expectMessage(
+            "warn: Not all groups specified in searchGroupBoosts were used in the documentation." +
+                " The unused groups were:\n\tGroupUnused"
+        );
+        logger.discardDebugMessages();
+        logger.expectNoOtherMessages();
+    },
 };