+++ src/test/converter2/behavior/searchCategoryBoosts.ts
@@ -1,0 +1,13 @@
+/**
+ * @category Cat1
+ * @category Cat2
+ */
+export interface A {}
+/**
+ * @category Cat0
+ */
+export interface B {}
+/**
+ * @category Cat1
+ */
+export interface C {}